# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Instagram-clone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/feed:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get home feed posts (from followed users)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Edit post caption
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /posts/{postId}/save:
    post:
      operationId: savePost
      tags: [posts]
      summary: Save / unsave a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResponse"
  /posts/upload:
    post:
      operationId: uploadPostMedia
      tags: [posts]
      summary: Upload media for a post (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadInput"
      responses:
        "200":
          description: Uploaded media URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadMediaResponse"

  # ── Comments ──────────────────────────────────────────────────────────────
  /posts/{postId}/comments:
    get:
      operationId: getPostComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /comments/{commentId}/like:
    post:
      operationId: likeComment
      tags: [comments]
      summary: Like/unlike a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /posts/{postId}/report:
    post:
      operationId: reportPost
      tags: [posts]
      summary: Report a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostReportInput"
      responses:
        "200":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── AI ────────────────────────────────────────────────────────────────────
...
  /push/subscribe:
    post:
      operationId: subscribeToPush
      tags: [notifications]
      summary: Register a browser push subscription for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionInput"
      responses:
        "201":
          description: Subscribed
  /push/unsubscribe:
    post:
      operationId: unsubscribeFromPush
      tags: [notifications]
      summary: Remove a browser push subscription for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                endpoint:
                  type: string
              required:
                - endpoint
      responses:
        "200":
          description: Unsubscribed

  # ── Conversations ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isFollowing:
          type: boolean
      required:
        - id
        - username
        - fullName
        - isFollowing
...
    Post:
      type: object
      properties:
        id:
          type: string
        caption:
          type: ["string", "null"]
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        audience:
          type: string
          enum: [everyone, close_friends]
        location:
          type: ["string", "null"]
        altText:
          type: ["string", "null"]
        commentsDisabled:
          type: boolean
        additionalMediaUrls:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        author:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
      required:
        - id
        - mediaUrl
        - mediaType
        - audience
        - commentsDisabled
        - additionalMediaUrls
        - likesCount
        - commentsCount
        - isLiked
        - isSaved
        - author
        - createdAt
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        hasMore:
          type: boolean
        total:
          type: integer
      required:
        - posts
        - hasMore
        - total
    PostInput:
      type: object
      properties:
        caption:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        audience:
          type: string
          enum: [everyone, close_friends]
        location:
          type: string
        altText:
          type: string
        commentsDisabled:
          type: boolean
        additionalMediaUrls:
          type: array
          items:
            type: string
      required:
        - mediaUrl
        - mediaType
    PostUpdate:
      type: object
      properties:
        caption:
          type: string
    LikeResponse:
      type: object
      properties:
        isLiked:
          type: boolean
        likesCount:
          type: integer
      required:
        - isLiked
        - likesCount
    SaveResponse:
      type: object
      properties:
        isSaved:
          type: boolean
      required:
        - isSaved

    # ── Comments ────────────────────────────────────────────────────────────
    Comment:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        author:
          $ref: "#/components/schemas/UserSummary"
        parentId:
          type: ["integer", "null"]
        likesCount:
          type: integer
        isLiked:
          type: boolean
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
      required:
        - id
        - text
        - author
        - likesCount
        - isLiked
        - replies
        - createdAt
    CommentInput:
      type: object
      properties:
        text:
          type: string
        parentId:
          type: string
      required:
        - text

    # ── Stories ─────────────────────────────────────────────────────────────
...
    PushSubscriptionInput:
      type: object
      properties:
        endpoint:
          type: string
        keys:
          type: object
          properties:
            p256dh:
              type: string
            auth:
              type: string
          required:
            - p256dh
            - auth
      required:
        - endpoint
        - keys

    # ── Messages ─────────────────────────────────────────────────────────────
...
    PostReportInput:
      type: object
      properties:
        reason:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    MediaUploadInput:
      type: object
      properties:
        data:
          type: string
          description: Base64 encoded file data
        mimeType:
          type: string
      required:
        - data
        - mimeType
    UploadMediaResponse:
      type: object
      properties:
        url:
          type: string
        publicId:
          type: string
        resourceType:
          type: string
      required:
        - url
        - publicId
        - resourceType
...